composer check fixed, Consumer moved
