+++ src/Graviton/MessageBundle/Command/ProduceCommand.php
@@ -1,5 +1,9 @@
 <?php
 
+/**
+ * Sends a given message to the message bus.
+ */
+
 namespace Graviton\MessageBundle\Command;
 
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
@@ -7,9 +11,21 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 
+/**
+ * Sends a given message to the message bus.
+ *
+ * @author   List of contributors <https://github.com/libgraviton/graviton/graphs/contributors>
+ * @license  http://opensource.org/licenses/gpl-license.php GNU Public License
+ * @link     http://swisscom.ch
+ */
 class ProduceCommand extends ContainerAwareCommand
 {
 
+    /**
+     * Configures command
+     *
+     * @return void
+     */
     protected function configure()
     {
         parent::configure();
@@ -30,9 +46,15 @@ class ProduceCommand extends ContainerAwareCommand
             );
     }
 
+    /**
+     * @param InputInterface  $input  The input
+     * @param OutputInterface $output The output
+     *
+     * @return void
+     */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $producer = $this->getContainer()->get('old_sound_rabbit_mq.job_producer');
+        $producer = $this->getContainer()->get('graviton.message.jobproducer');
         $producer->setContentType('application/json');
         $producer->publish(
             $input->getArgument('data'),
