+++ src/Graviton/MessageBundle/Resources/config/services.xml
@@ -4,7 +4,7 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <parameters>
         <parameter key="graviton.message.documenteventpublisher.class">Graviton\MessageBundle\Service\DocumentEventPublisher</parameter>
-        <parameter key="graviton.message.consumer.dump.class">Graviton\MessageBundle\Consumer\DumpConsumer</parameter>
+        <parameter key="graviton.message.consumer.dump.class">Graviton\MessageBundle\Service\DumpConsumer</parameter>
         <parameter key="graviton.rest.routing.collection.class">Symfony\Component\Routing\RouteCollection</parameter>
     </parameters>
     <services>
